apps/backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
gunicorn==21.2.0
packaging==23.2
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
whitenoise==6.6.0
